% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/strata.R
\name{check_pop_levels}
\alias{check_pop_levels}
\title{Check the use of pop.levels, pop.labels and pop.select arguments.}
\usage{
check_pop_levels(pop.levels = NULL, pop.labels = NULL, pop.select = NULL)
}
\arguments{
\item{pop.levels}{(optional, string) This refers to the levels in a factor. In this
case, the id of the pop.
Use this argument to have the pop ordered your way instead of the default
alphabetical or numerical order. e.g. \code{pop.levels = c("QUE", "ONT", "ALB")}
instead of the default \code{pop.levels = c("ALB", "ONT", "QUE")}.
White spaces in population names are replaced by underscore.
Default: \code{pop.levels = NULL}.}

\item{pop.labels}{(optional, string) Use this argument to rename/relabel
your pop or combine your pop. e.g. To combine \code{"QUE"} and \code{"ONT"}
into a new pop called \code{"NEW"}:
(1) First, define the levels for your pop with \code{pop.levels} argument:
\code{pop.levels = c("QUE", "ONT", "ALB")}.
(2) then, use \code{pop.labels} argument:
\code{pop.labels = c("NEW", "NEW", "ALB")}.
To rename \code{"QUE"} to \code{"TAS"}:
\code{pop.labels = c("TAS", "ONT", "ALB")}.
Default: \code{pop.labels = NULL}.
White spaces in population names are replaced by underscore.}

\item{pop.select}{(optional, string) Selected list of populations for
the analysis. e.g. \code{pop.select = c("QUE", "ONT")} to select \code{QUE}
and \code{ONT} population samples (out of 20 pops). If \code{pop.labels}
argument was used to rename the strata column, use the new names with
\code{pop.select}.
Default: \code{pop.select = NULL}.}
}
\description{
Check that pop.levels and pop.labels and pop.select arguments
are used correctly and that the values are cleaned for spaces.
}
\author{
Thierry Gosselin \email{thierrygosselin@icloud.com}
}
