% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dart.R
\name{checks_dart_target}
\alias{checks_dart_target}
\title{checks_dart_target}
\usage{
checks_dart_target(target.id, strata, path.folder = NULL, verbose = TRUE)
}
\arguments{
\item{path.folder}{Default: \code{path.folder = NULL}.}

\item{verbose}{Default: \code{verbose = TRUE}.}
}
\description{
Checks target.id and strata for DArT potential problems.
}
\keyword{internal}
