% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{clean_ind_names}
\alias{clean_ind_names}
\title{Clean individual's names for genomic workflows}
\usage{
clean_ind_names(x)
}
\arguments{
\item{x}{(character string) Individuals character string.}
}
\description{
Function to clean individual's name
that interfere with some packages
or codes. \code{_} and \code{:} are changed to a dash \code{-}.
Whitespaces are removed.
Used internally in \href{https://github.com/thierrygosselin/genometranslator}{genometranslator}
and might be of interest for users.
}
