% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dart.R
\name{clean_normalize_dart_metadata}
\alias{clean_normalize_dart_metadata}
\title{clean_normalize_dart_metadata}
\usage{
clean_normalize_dart_metadata(
  x,
  dart.check,
  blacklist.id = NULL,
  strata = NULL
)
}
\arguments{
\item{blacklist.id}{Default: \code{blacklist.id = NULL}.}

\item{strata}{Default: \code{strata = NULL}.}
}
\description{
clean_normalize_dart_metadata (only the DArT columns = snakecase...)
}
\keyword{internal}
