% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dart.R
\name{dart_genotyping_strategy}
\alias{dart_genotyping_strategy}
\title{dart_genotyping_strategy}
\usage{
dart_genotyping_strategy(
  n.markers,
  gt.vcf = NULL,
  gt.vcf.nuc = NULL,
  gt = NULL
)
}
\arguments{
\item{gt.vcf}{Default: \code{gt.vcf = NULL}.}

\item{gt.vcf.nuc}{Default: \code{gt.vcf.nuc = NULL}.}

\item{gt}{Default: \code{gt = NULL}.}
}
\description{
Generating genotypes coding is computer costly, strategy ...
}
\keyword{internal}
