% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dart.R
\name{detect_calibration_problem}
\alias{detect_calibration_problem}
\title{detect_calibration_problem}
\usage{
detect_calibration_problem(x, dart.check, verbose = TRUE)
}
\arguments{
\item{verbose}{Logical. Display calibration messages.
Default: \code{verbose = TRUE}.}
}
\description{
Detect DArT with REF and ALT not calibrated... based on alleles or depth
}
\keyword{internal}
