% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/detect_genomic_format.R
\name{detect_dart_format}
\alias{detect_dart_format}
\title{detect_dart_format}
\usage{
detect_dart_format(data, verbose = TRUE)
}
\arguments{
\item{verbose}{Default: \code{verbose = TRUE}.}
}
\description{
Detect the dart genotype format: 1row, 2rows or counts
}
\keyword{internal}
