% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/detect_genomic_format.R
\name{detect_genomic_format}
\alias{detect_genomic_format}
\title{Used internally in radiator to detect the file format}
\usage{
detect_genomic_format(data, guess = NULL)
}
\arguments{
\item{data}{15 options for input: VCFs (SNPs or Haplotypes,
to make the vcf population ready),
PLINK 1 (bed, ped, and tped), PLINK 2 (pgen), stacks haplotype file,
genind (library(adegenet)),
genlight (library(adegenet)), gtypes (library(strataG)), genepop, DArT,
and a data frame in long/tidy or wide format. To verify that radiator detect
your file format use \code{\link{detect_genomic_format}} (see example below).
New addition to radiator: the Apache Parquet columnar storage file format that
will replace the fst (Lightning Fast Serialiation) format.
Documented in \strong{Input genomic datasets} of \code{\link{tidy_genome}}.}

\item{guess}{(character) In development, guess faster the type of file.
Default: \code{guess = NULL}.}
}
\value{
One of these file format:
\itemize{
\item tbl_df: for a data frame
\item genind: for a genind object
\item genlight: for a genlight object
\item gtypes: for a gtypes object
\item vcf.file: for a vcf file
\item plink.tped.file: for a plink tped file
\item plink.bed.file: for a plink bed file
\item plink.ped.file: for a plink ped file
\item plink.pgen.file: for a PLINK 2 pgen file
\item genepop.file: for a genepop file
\item haplo.file: for a stacks haplotypes file
\item fstat.file: for a fstat file
\item dart: for a DArT file
\item fst.file: for a file ending with .rad
\item SeqVarGDSClass: for SeqArray GDS file.
\item arrow parquet: for a Apache Parquet columnar storage file format,
used by arrow R package.
}
}
\description{
Detect the file format of genomic data set.
}
\examples{
\dontrun{
#To verify your file is detected by radiator as the correct format:
genometranslator::detect_genomic_format(data = "populations.snps.vcf")
}
}
\author{
Thierry Gosselin \email{thierrygosselin@icloud.com}
}
