% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genotype_encoding.R
\name{detect_gt}
\alias{detect_gt}
\title{detect_gt}
\usage{
detect_gt(
  x,
  gt.format = c("GT", "ALT_DOSAGE", "GT_VCF", "GT_VCF_NUC"),
  keep.one = TRUE,
  favorite = "ALT_DOSAGE"
)
}
\arguments{
\item{x}{The data}

\item{gt.format}{(character)
Default: \code{gt.format = c("GT", "ALT_DOSAGE", "GT_VCF", "GT_VCF_NUC")}.}

\item{keep.one}{(logical) Will return only one format if \code{keep.one = TRUE}.
Default: \code{keep.one = TRUE}.}

\item{favorite}{If more than one format is present and \code{keep.one = TRUE},
the favorite will be returned, if present. Otherwise, the first format in
\code{gt.format} is returned.
Default: \code{favorite = "ALT_DOSAGE"}.}
}
\description{
Detect the genotype format used in the data set.
}
\keyword{internal}
