% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vcf.R
\name{dots_keepers_read_vcf}
\alias{dots_keepers_read_vcf}
\title{Dot-keepers for read_vcf}
\usage{
dots_keepers_read_vcf(extra = NULL, exclude = NULL)
}
\arguments{
\item{extra}{Default: \code{extra = NULL}.}

\item{exclude}{Default: \code{exclude = NULL}.}
}
\description{
Return the set of \code{...} arguments supported by \code{read_vcf()}.
This builds on the global core keepers and adds/removes any
read_vcf-specific arguments.
}
\keyword{internal}
