% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gds.R
\name{check_coverage}
\alias{check_coverage}
\alias{extract_coverage}
\title{check_coverage}
\usage{
check_coverage(
  gds,
  genotypes.metadata.check = FALSE,
  stacks.haplo.check = FALSE,
  dart.check = FALSE
)

extract_coverage(
  gds,
  individuals = TRUE,
  markers = TRUE,
  coverage = TRUE,
  allele.coverage = TRUE,
  coverage.stats = c("sum", "mean", "median", "iqr"),
  subsample.info = 1,
  verbose = TRUE,
  parallel.core = parallel::detectCores() - 1
)
}
\arguments{
\item{gds}{The gds object.}

\item{genotypes.metadata.check}{(optional, logical) Look for already extracted coverage information in the
radiator genotypes_metadata field of the GDS.
Default: \code{genotypes.metadata.check = FALSE}.}

\item{stacks.haplo.check}{(optional, logical) stacks haplotypes VCF header is
baddly generated. It will say you have Read and allele Depth info, but you don't.
Default: \code{stacks.haplo.check = FALSE}.}

\item{dart.check}{(optional, logical) DArT have different reporting for coverage
information. Will no longer report the average coverage stats from 1 and 2-rows DArT format.
Default: \code{dart.check = FALSE}.}

\item{individuals}{(optional, logical) Default: \code{individuals = TRUE}.}

\item{markers}{(optional, logical) Default: \code{markers = TRUE}.}

\item{coverage}{(optional, logical) Default: \code{coverage = TRUE}.}

\item{allele.coverage}{(optional, logical) Default: \code{ad = TRUE}.
Default: \code{allele.coverage = TRUE}.}

\item{coverage.stats}{(optional, character string). Choice of stats to use with
coverage.
Default: \code{coverage.stats = c("sum", "mean", "median", "iqr")}.}

\item{subsample.info}{(optional, double) Default: \code{subsample.info = 1}.
The subsample proportion used (e.g. 0.3 or none the default).}

\item{verbose}{Logical indicating whether progress messages are emitted.
Default: \code{verbose = TRUE}.}

\item{parallel.core}{Number of workers available for parallel operations.
Default: \code{parallel.core = parallel::detectCores() - 1}.}
}
\description{
Check that the coverage info is in the GDS. By default, it will
look for the DP and AD info in the FORMAT field.

Extract coverage information from a GDS file
}
