% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gds.R
\name{extract_genotypes_metadata}
\alias{extract_genotypes_metadata}
\title{extract_genotypes_metadata}
\usage{
extract_genotypes_metadata(
  gds,
  genotypes.meta.select = NULL,
  genotypes = FALSE,
  metadata.node = TRUE,
  index.only = FALSE,
  sync.markers.individuals = TRUE,
  whitelist = FALSE,
  blacklist = FALSE,
  verbose = FALSE
)
}
\arguments{
\item{gds}{The gds object.}

\item{genotypes.meta.select}{(optional, character) Default:\code{genotypes.meta.select = NULL}.
Default: \code{genotypes.meta.select = NULL}.}

\item{genotypes}{(optional, character) Default: \code{genotypes = FALSE}.}

\item{metadata.node}{(optional, logical) Default: \code{metadata.node = TRUE}.}

\item{index.only}{(optional, logical) Default: \code{index.only = FALSE}.}

\item{sync.markers.individuals}{(optional, logical) Default: \code{sync.markers.individuals = TRUE}.}

\item{whitelist}{(optional, logical) Default: \code{whitelist = FALSE}.}

\item{blacklist}{(optional, logical) Default: \code{blacklist = FALSE}.}

\item{verbose}{Logical indicating whether progress messages are emitted.
Default: \code{verbose = FALSE}.}
}
\description{
Import gds or radiator genotypes meta node
}
