% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gds.R
\name{extract_individuals_metadata}
\alias{extract_individuals_metadata}
\title{extract_individuals_metadata}
\usage{
extract_individuals_metadata(
  gds,
  ind.field.select = NULL,
  metadata.node = TRUE,
  whitelist = FALSE,
  blacklist = FALSE,
  verbose = FALSE
)
}
\arguments{
\item{gds}{The gds object.}

\item{ind.field.select}{(optional, character) Default:\code{ind.field.select = NULL}.
Default: \code{ind.field.select = NULL}.}

\item{metadata.node}{(optional, logical) Default:\code{metadata.node = TRUE}.
Default: \code{metadata.node = TRUE}.}

\item{whitelist}{(optional, logical) Default:\code{whitelist = FALSE}.
Default: \code{whitelist = FALSE}.}

\item{blacklist}{(optional, logical) Default:\code{blacklist = FALSE}.
Default: \code{blacklist = FALSE}.}

\item{verbose}{Logical indicating whether progress messages are emitted.
Default: \code{verbose = FALSE}.}
}
\description{
Import individuals metadata from a package or SeqArray node
}
