% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vcf.R
\name{filter_haplotype_format}
\alias{filter_haplotype_format}
\title{Filter non-biallelic / haplotype-style variants in a radiator GDS}
\usage{
filter_haplotype_format(
  gds,
  filters.parameters,
  path.folder,
  file.date,
  filter.haplotype.format = FALSE,
  internal = TRUE,
  verbose = TRUE
)
}
\arguments{
\item{gds}{A radiator GDS object (\code{SeqVarGDSClass}) with a \code{/radiator/markers.meta}
node containing at least \code{MARKERS}, \code{REF} and \code{ALT}.}

\item{filters.parameters}{The current filter-parameter object generated by \code{genome_parameters()}.}

\item{path.folder}{(character)
Root folder used to store filter outputs.}

\item{file.date}{(character)
Timestamp string (e.g. \code{"20251203@1919"}) used in filenames.}

\item{filter.haplotype.format}{(logical)
Activate or deactivate this filter.
Default: \code{filter.haplotype.format = FALSE}.}

\item{internal}{(logical)
Indicate if the function is called internally by other radiator functions,
used for controlling side-effect messages and paths.
Default: \code{internal = TRUE}.}

\item{verbose}{(logical)
When \code{TRUE}, the function prints messages about detected haplotypes and
written files.
Default: \code{verbose = TRUE}.}
}
\value{
A list with:
\itemize{
\item \code{markers.meta} – updated markers metadata tibble;
\item \code{filters.parameters} – updated filter-parameter object.
}
}
\description{
Filter non-biallelic / haplotype-style variants in a radiator GDS
}
