% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vcf.R
\name{filter_vcf_filter_column}
\alias{filter_vcf_filter_column}
\title{Filter markers based on the VCF FILTER column}
\usage{
filter_vcf_filter_column(
  gds,
  markers.meta,
  filters.parameters,
  path.folder,
  file.date,
  verbose = TRUE
)
}
\arguments{
\item{gds}{A SeqArray GDS object containing the imported VCF.}

\item{markers.meta}{The current markers metadata tibble (from radiator).}

\item{filters.parameters}{The filters.parameters object to update.}

\item{path.folder}{Path to the filtering results folder.}

\item{file.date}{Character string used for naming result files.}

\item{verbose}{(logical) Verbosity.

Default: \code{verbose = TRUE}.}
}
\value{
A list with:
\itemize{
\item \code{markers.meta}: updated tibble;
\item \code{filters.parameters}: updated parameters object.
}
}
\description{
Filter markers based on the VCF FILTER column
}
\keyword{internal}
