% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/marker_metadata.R
\name{generate_markers_metadata}
\alias{generate_markers_metadata}
\title{Generate markers metadata}
\usage{
generate_markers_metadata(
  data,
  generate.markers.metadata = TRUE,
  generate.ref.alt = FALSE,
  biallelic = NULL,
  parallel.core = parallel::detectCores() - 1,
  verbose = TRUE
)
}
\arguments{
\item{data}{An object with a column named \code{MARKERS}.
If \code{CHROM}, \code{LOCUS}, \code{POS} are already present, the function
returns the dataset untouched.
The data can be whitelists and blacklists of markers or tidy datasets or
genome GDS object.}

\item{generate.markers.metadata}{(logical, optional)
Generate missing markers metadata when missing.
\code{"CHROM", "LOCUS", "POS"}.
Default: \code{generate.markers.metadata = TRUE}.}

\item{generate.ref.alt}{(logical, optional) Generate missing REF/ALT alleles
with: REF = A and ALT = C (for biallelic datasets, only).
It is turned off automatically
when argument \code{markers.meta.lists.only = TRUE} and
on automatically when argument \code{markers.meta.all.only = TRUE}
Default: \code{generate.ref.alt = FALSE}.}

\item{biallelic}{(logical) Speed up the function execution by entering
if the dataset is biallelic or not. Used internally for verification, before
generating REF/ALT info.
The argument is required when \code{generate.ref.alt = TRUE} and REF/ALT
information is absent. Biallelic detection is analysis policy and is not
performed automatically by this package.
Default: \code{biallelic = NULL}.}

\item{parallel.core}{Default: \code{parallel.core = parallel::detectCores() - 1}.}
}
\value{
Depending on argument's value, the same data is returned
in the global environment, with potential these additional columns:
\code{CHROM, LOCUS, POS, REF, ALT}.
}
\description{
Generate markers metadata: \code{CHROM, LOCUS, POS, REF, ALT}
when missing from tidy datasets.
}
\examples{
\dontrun{
tidy.data <- genometranslator::generate_markers_metadata(data = bluefintuna.data)
}
}
\seealso{
\code{\link{separate_markers}}
}
\author{
Thierry Gosselin \email{thierrygosselin@icloud.com}
}
\keyword{internal}
