% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gds.R
\name{generate_stats}
\alias{generate_stats}
\title{generate_stats}
\usage{
generate_stats(
  gds,
  individuals = TRUE,
  markers = TRUE,
  missing = TRUE,
  heterozygosity = TRUE,
  coverage = TRUE,
  allele.coverage = TRUE,
  mac = TRUE,
  snp.position.read = TRUE,
  snp.per.locus = TRUE,
  subsample = NULL,
  exhaustive = TRUE,
  force.stats = TRUE,
  path.folder = NULL,
  plot = TRUE,
  digits = 6,
  file.date = NULL,
  parallel.core = parallel::detectCores() - 1,
  verbose = TRUE
)
}
\arguments{
\item{individuals}{Default: \code{individuals = TRUE}.}

\item{markers}{Default: \code{markers = TRUE}.}

\item{missing}{Default: \code{missing = TRUE}.}

\item{heterozygosity}{Default: \code{heterozygosity = TRUE}.}

\item{coverage}{Default: \code{coverage = TRUE}.}

\item{allele.coverage}{Default: \code{allele.coverage = TRUE}.}

\item{mac}{Default: \code{mac = TRUE}.}

\item{snp.position.read}{Default: \code{snp.position.read = TRUE}.}

\item{snp.per.locus}{Default: \code{snp.per.locus = TRUE}.}

\item{subsample}{Default: \code{subsample = NULL}.}

\item{exhaustive}{Default: \code{exhaustive = TRUE}.}

\item{force.stats}{Default: \code{force.stats = TRUE}.}

\item{path.folder}{Default: \code{path.folder = NULL}.}

\item{plot}{Default: \code{plot = TRUE}.}

\item{digits}{Default: \code{digits = 6}.}

\item{file.date}{Default: \code{file.date = NULL}.}

\item{parallel.core}{Default: \code{parallel.core = parallel::detectCores() - 1}.}

\item{verbose}{Default: \code{verbose = TRUE}.}
}
\description{
Generate individuals and markers statistics
}
\keyword{internal}
