% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gds.R
\name{genome_gds}
\alias{genome_gds}
\title{Genome GDS constructor}
\usage{
genome_gds(
  genotypes = NULL,
  biallelic = TRUE,
  data.source = NULL,
  strata = NULL,
  markers.meta = NULL,
  genotypes.meta = NULL,
  dp = NULL,
  ad = NULL,
  gl = NULL,
  filename = NULL,
  open = FALSE,
  verbose = TRUE
)
}
\arguments{
\item{genotypes}{Default: \code{genotypes = NULL}.}

\item{biallelic}{Default: \code{biallelic = TRUE}.}

\item{data.source}{Default: \code{data.source = NULL}.}

\item{strata}{Default: \code{strata = NULL}.}

\item{markers.meta}{Default: \code{markers.meta = NULL}.}

\item{genotypes.meta}{Default: \code{genotypes.meta = NULL}.}

\item{dp}{Default: \code{dp = NULL}.}

\item{ad}{Default: \code{ad = NULL}.}

\item{gl}{Default: \code{gl = NULL}.}

\item{filename}{Default: \code{filename = NULL}.}

\item{open}{Default: \code{open = FALSE}.}

\item{verbose}{Default: \code{verbose = TRUE}.}
}
\description{
helper function to construct a genome GDS
}
\keyword{internal}
