% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genome_parameters.R
\name{genome_info}
\alias{genome_info}
\title{Summarise genomic data dimensions}
\usage{
genome_info(x, verbose = FALSE)
}
\arguments{
\item{x}{A tidy genomic data frame or supported GDS object.}

\item{verbose}{Display the summary.

Default: \code{verbose = FALSE}.}
}
\value{
A list containing counts of chromosomes, loci, markers, strata, and
individuals.
}
\description{
Summarise genomic data dimensions
}
