% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gds.R
\name{gt2array}
\alias{gt2array}
\title{gt2array}
\usage{
gt2array(genotypes, n.ind = NULL, n.snp = NULL)
}
\arguments{
\item{n.ind}{Default: \code{n.ind = NULL}.}

\item{n.snp}{Default: \code{n.snp = NULL}.}
}
\description{
Alternate allele dosage gt (ALT_DOSAGE) to
presence/absence array for the genotypes in GDS
}
\keyword{internal}
