% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genotype_encoding.R
\name{gt_recoding}
\alias{gt_recoding}
\title{gt_recoding}
\usage{
gt_recoding(
  x,
  gt = TRUE,
  alt.dosage = TRUE,
  gt.vcf = TRUE,
  gt.vcf.nuc = TRUE,
  arrange = TRUE
)
}
\arguments{
\item{gt}{Default: \code{gt = TRUE}.}

\item{alt.dosage}{Default: \code{alt.dosage = TRUE}.}

\item{gt.vcf}{Default: \code{gt.vcf = TRUE}.}

\item{gt.vcf.nuc}{Default: \code{gt.vcf.nuc = TRUE}.}

\item{arrange}{Default: \code{arrange = TRUE}.}
}
\description{
Translate among supported genotype encodings.
}
\keyword{internal}
