% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gds.R
\name{id_qc_helper}
\alias{id_qc_helper}
\title{id_qc_helper}
\usage{
id_qc_helper(
  x,
  x.sum,
  stats.id = c("missing", "heterozygosity", "coverage"),
  path.folder = NULL
)
}
\arguments{
\item{stats.id}{Default: \code{stats.id = c("missing", "heterozygosity", "coverage")}.}

\item{path.folder}{Default: \code{path.folder = NULL}.}
}
\description{
Generate helpers tables and blacklists for individual's QC
}
\keyword{internal}
