% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gds.R
\name{markers_het}
\alias{markers_het}
\title{markers_het}
\usage{
markers_het(gds, parallel.core = TRUE)
}
\arguments{
\item{parallel.core}{Default: \code{parallel.core = TRUE}.}
}
\description{
Calculate markers het
}
\keyword{internal}
