% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gds.R
\name{parse_gds_metadata}
\alias{parse_gds_metadata}
\title{parse_gds_metadata}
\usage{
parse_gds_metadata(x, gds = NULL, verbose = TRUE, strip.rad = FALSE)
}
\arguments{
\item{gds}{Default: \code{gds = NULL}.}

\item{verbose}{Default: \code{verbose = TRUE}.}

\item{strip.rad}{Default: \code{strip.rad = FALSE}.}
}
\description{
function to parse the format field and tidy the results of VCF
}
\keyword{internal}
