% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genotype_encoding.R
\name{separate_gt}
\alias{separate_gt}
\title{separate_gt}
\usage{
separate_gt(
  x,
  gt = "GT_VCF_NUC",
  gather = TRUE,
  haplotypes = FALSE,
  exclude = c("LOCUS", "INDIVIDUALS", "POP_ID"),
  alleles.naming = c("A1", "A2"),
  remove = TRUE,
  filter.missing = FALSE,
  split.chunks = 3,
  parallel.core = parallel::detectCores() - 1
)
}
\arguments{
\item{gt}{Default: \code{gt = "GT_VCF_NUC"}.}

\item{gather}{Default: \code{gather = TRUE}.}

\item{haplotypes}{Default: \code{haplotypes = FALSE}.}

\item{exclude}{Default: \code{exclude = c("LOCUS", "INDIVIDUALS", "POP_ID")}.}

\item{alleles.naming}{Default: \code{alleles.naming = c("A1", "A2")}.}

\item{remove}{Default: \code{remove = TRUE}.}

\item{filter.missing}{Default: \code{filter.missing = FALSE}.}

\item{split.chunks}{Default: \code{split.chunks = 3}.}

\item{parallel.core}{Default: \code{parallel.core = parallel::detectCores() - 1}.}
}
\description{
Separate genotype field
}
\keyword{internal}
