% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/strata.R
\name{strata_haplo}
\alias{strata_haplo}
\title{strata_haplo}
\usage{
strata_haplo(strata = NULL, data = NULL, blacklist.id = NULL)
}
\arguments{
\item{strata}{Default: \code{strata = NULL}.}

\item{data}{Default: \code{data = NULL}.}

\item{blacklist.id}{Default: \code{blacklist.id = NULL}.}
}
\description{
Manage strata
}
\keyword{internal}
