% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gds.R
\name{summary_gds}
\alias{summary_gds}
\title{summary_gds}
\usage{
summary_gds(gds, check.sync = FALSE, verbose = TRUE)
}
\arguments{
\item{gds}{A genome GDS file path or object.
Default: \code{gds = NULL}.}

\item{check.sync}{(logical, optional) Check if GDS object is in sync with
radiator metada.
Default: \code{check.sync = FALSE}.}

\item{verbose}{Logical indicating whether progress messages are emitted.
Default: \code{verbose = TRUE}.}
}
\description{
Summary of gds object or file: number of samples and markers
}
