% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gds.R
\name{update_bl_markers}
\alias{update_bl_markers}
\title{update_bl_markers}
\usage{
update_bl_markers(
  gds,
  generate = FALSE,
  update = NULL,
  extract = FALSE,
  bl.gds = NULL
)
}
\arguments{
\item{generate}{Default: \code{generate = FALSE}.}

\item{update}{Default: \code{update = NULL}.}

\item{extract}{Default: \code{extract = FALSE}.}

\item{bl.gds}{Default: \code{bl.gds = NULL}.}
}
\description{
Generate and update blacklist of markers on radiator GDS node
}
\keyword{internal}
