% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gds.R
\name{update_genome_gds}
\alias{update_genome_gds}
\title{update_genome_gds}
\usage{
update_genome_gds(
  gds,
  metadata.node = TRUE,
  node.name,
  value,
  replace = TRUE,
  sync = FALSE,
  summary = FALSE,
  verbose = FALSE
)
}
\arguments{
\item{metadata.node}{Default: \code{metadata.node = TRUE}.}

\item{replace}{Default: \code{replace = TRUE}.}

\item{sync}{Default: \code{sync = FALSE}.}

\item{summary}{Default: \code{summary = FALSE}.}

\item{verbose}{Default: \code{verbose = FALSE}.}
}
\description{
Update a genome GDS metadata node
}
\keyword{internal}
