% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vcf.R
\name{vcf_strata}
\alias{vcf_strata}
\title{Join stratification metadata to a VCF (population-aware VCF)}
\usage{
vcf_strata(data, strata, filename = NULL)
}
\arguments{
\item{data}{A VCF file}

\item{strata}{(optional) A tab delimited file at least 2 columns
with header:
\code{INDIVIDUALS} and \code{STRATA}.
The \code{STRATA} and any other columns can be any hierarchical grouping.
Supply a sample metadata table as described in \code{\link[=read_strata]{read_strata()}}.}

\item{filename}{(optional) The file name for the modifed VCF,
written to the working directory. Default: \code{filename = NULL} will make a
custom filename with data and time.
Default: \code{filename = NULL}.}
}
\value{
A VCF file in the working directory with new \code{FORMAT} field(s)
correponding to the strata column(s).
}
\description{
Include stratification metadata, e.g. population-level information,
to the \code{FORMAT} field of a VCF file.
}
\references{
Danecek P, Auton A, Abecasis G et al. (2011)
The variant call format and VCFtools.
Bioinformatics, 27, 2156-2158.
}
\seealso{
\href{https://vcftools.github.io}{VCF web page}

\href{VCF specification page}{https://vcftools.github.io/specs.html}
}
\author{
Thierry Gosselin \email{thierrygosselin@icloud.com}
}
\keyword{internal}
