% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_gsi_sim.R
\name{write_gsi_sim}
\alias{write_gsi_sim}
\title{Write a gsi_sim file from a data frame (wide or long/tidy).}
\usage{
write_gsi_sim(
  data,
  pop.levels = NULL,
  pop.labels = NULL,
  strata = NULL,
  filename = "gsi_sim.unname.txt"
)
}
\arguments{
\item{data}{A tidy genomic data set in the working directory tidy formats.
\emph{How to get a tidy data frame ?}
Look for \pkg{genometranslator} \code{\link{tidy_genome}}.}

\item{pop.levels}{(option, string) This refers to the levels in a factor. In this
case, the id of the pop.
Use this argument to have the pop ordered your way instead of the default
alphabetical or numerical order. e.g. \code{pop.levels = c("QUE", "ONT", "ALB")}
instead of the default \code{pop.levels = c("ALB", "ONT", "QUE")}.
Default: \code{pop.levels = NULL}. If you find this too complicated, there is also the
\code{strata} argument that can do the same thing, see below.}

\item{pop.labels}{(optional, string) Use this argument to rename/relabel
your pop or combine your pop. e.g. To combine \code{"QUE"} and \code{"ONT"}
into a new pop called \code{"NEW"}:
(1) First, define the levels for your pop with \code{pop.levels} argument:
\code{pop.levels = c("QUE", "ONT", "ALB")}.
(2) then, use \code{pop.labels} argument:
\code{pop.levels = c("NEW", "NEW", "ALB")}.#'
To rename \code{"QUE"} to \code{"TAS"}:
\code{pop.labels = c("TAS", "ONT", "ALB")}.
Default: \code{pop.labels = NULL}. If you find this too complicated, there is also the
\code{strata} argument that can do the same thing, see below.}

\item{strata}{(optional) A tab delimited file with 2 columns with header:
\code{INDIVIDUALS} and \code{STRATA}.
Default: \code{strata = NULL}. Use this argument to rename or change
the populations id with the new \code{STRATA} column.
The \code{STRATA} column can be any hierarchical grouping.}

\item{filename}{The name of the file written to the working directory.
Default: \code{filename = "gsi_sim.unname.txt"}.}
}
\value{
A gsi_sim input file is saved to the working directory.
}
\description{
Write a gsi_sim file from a data frame (wide or long/tidy).
Used internally in \href{https://github.com/thierrygosselin/assigner}{assigner} and
\href{https://github.com/thierrygosselin/genometranslator}{genometranslator}
and might be of interest for users.
}
\section{Data filtering}{

This writer does not silently filter markers or individuals. It may validate
requirements imposed by the destination format and stop with an informative
error when the input is unsuitable. It is the user's responsibility to
filter and quality-control the data appropriately for the intended analysis
before generating the output. Use
\href{https://thierrygosselin.github.io/radr/}{\pkg{radr}} or another
suitable workflow when filtering is required.
}

\section{Dependencies}{

Required package dependencies are declared in \code{DESCRIPTION} and are
installed with \pkg{genometranslator}. Any additional dependency needed only
for this format or option is identified in this help page. Use
\code{genometranslator_dependencies()} to inspect the availability of core
packages, optional packages, and external executables.
}

\references{
Anderson, Eric C., Robin S. Waples, and Steven T. Kalinowski. (2008)
An improved method for predicting the accuracy of genetic stock identification.
Canadian Journal of Fisheries and Aquatic Sciences 65, 7:1475-1486.

Anderson, E. C. (2010) Assessing the power of informative subsets of
loci for population assignment: standard methods are upwardly biased.
Molecular ecology resources 10, 4:701-710.
}
\seealso{
\href{https://github.com/eriqande/gsi_sim}{gsi_sim} and
\href{https://github.com/eriqande/rubias}{rubias}: genetic stock
identification (GSI) in the tidyverse.
}
\author{
Thierry Gosselin \email{thierrygosselin@icloud.com}
}
