% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calibrate_alleles.R
\name{calibrate_alleles}
\alias{calibrate_alleles}
\title{Calibrate REF and ALT alleles based on count}
\usage{
calibrate_alleles(
  data,
  biallelic = NULL,
  parallel.core = parallel::detectCores() - 1,
  verbose = FALSE,
  ...
)
}
\arguments{
\item{data}{A genomic data set in the global environment tidy formats.
See details for more info.}

\item{biallelic}{(optional) If \code{biallelic = TRUE/FALSE} will be use
during multiallelic REF/ALT decision and speed up computations.
Default: \code{biallelic = NULL}.}

\item{parallel.core}{(optional) The number of core used for parallel
execution. This is no longer used. The code is as fast as it can. Using
more cores will reduce the speed.
Default: \code{parallel.core = parallel::detectCores() - 1}.}

\item{verbose}{(optional, logical) \code{verbose = TRUE} to be chatty
during execution.
Default: \code{verbose = FALSE}.}

\item{...}{(optional) To pass further argument for fine-tuning the tidying
(details below).}
}
\value{
Depending if the input file is biallelic or multiallelic,
the function will output additional to REF and ALT column several genotype codings:
\itemize{
\item \code{GT}: the genotype in 6 digits format with integers.
\item \code{GT_VCF}: the genotype in VCF format with integers.
\item \code{GT_VCF_NUC}: the genotype in VCF format with letters corresponding to nucleotide.
\item \code{ALT_DOSAGE}: alternate-allele dosage for biallelic markers:
\code{0}, \code{1}, or \code{2} copies of the ALT allele, and \code{NA}
for missing genotypes.
}
}
\description{
Calibrate REF and ALT alleles based on counts. The REF allele is
designated as the allele with more counts in the dataset.
The function will generate a REF and ALT columns.

\strong{reference genome}: for people using a reference genome, the reference
allele terminology is different and is not based on counts...

Used internally in \pkg{genometranslator}
and might be of interest for users.
}
\details{
\strong{Input data:}
A minimum of 4 columns are required (the rest are considered metata info):
\enumerate{
\item \code{MARKERS}
\item \code{POP_ID}
\item \code{INDIVIDUALS}
\item \code{GT} and/or \code{GT_VCF_NUC} and/or \code{GT_VCF}
}

\emph{How to get a tidy data frame ?}
\pkg{genometranslator} \code{\link{tidy_genome}}
}
\author{
Thierry Gosselin \email{thierrygosselin@icloud.com}
}
