% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dart.R
\name{dart_genotyper}
\alias{dart_genotyper}
\title{dart_genotyper}
\usage{
dart_genotyper(
  x,
  dart.check,
  markers.metadata = NULL,
  gt.vcf = FALSE,
  gt.vcf.nuc = FALSE,
  gt = FALSE
)
}
\arguments{
\item{markers.metadata}{Default: \code{markers.metadata = NULL}.}

\item{gt.vcf}{Default: \code{gt.vcf = FALSE}.}

\item{gt.vcf.nuc}{Default: \code{gt.vcf.nuc = FALSE}.}

\item{gt}{Default: \code{gt = FALSE}.}
}
\description{
Normalization of DArT genotypes
}
\keyword{internal}
