% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gds.R
\name{list_filters}
\alias{list_filters}
\title{List current active filters in a genometranslator GDS object}
\usage{
list_filters(gds)
}
\arguments{
\item{gds}{A genome GDS file path or object.
Default: \code{gds = NULL}.}
}
\description{
List current active filters for individuals and markers in a
genometranslator GDS object.
}
\examples{
\dontrun{
# List active filters for individuals and markers
list_filters(gds)
}
}
\seealso{
\code{\link{sync_gds}}, \code{\link{list_filters}}.
}
\author{
Thierry Gosselin \email{thierrygosselin@icloud.com}
}
