% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gds.R
\name{missing_per_pop}
\alias{missing_per_pop}
\title{missing_per_pop}
\usage{
missing_per_pop(gds, strata, parallel.core = parallel::detectCores() - 1)
}
\arguments{
\item{parallel.core}{Default: \code{parallel.core = parallel::detectCores() - 1}.}
}
\description{
Generate missingness per markers per pop helper table
}
\keyword{internal}
