% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/strata.R
\name{read_blacklist_id}
\alias{read_blacklist_id}
\title{read_blacklist_id}
\usage{
read_blacklist_id(blacklist.id = NULL, verbose = TRUE)
}
\arguments{
\item{blacklist.id}{(optional, path or object) A blacklist file in the working directory
or object in the global environment. The data frame
as 1 column (named \code{INDIVIDUALS}) and is filled with the individual IDs
The ids are cleaned with \code{\link{clean_ind_names}} for separators,
only \code{-} are tolerated. Duplicates are removed automatically.
Default: \code{blacklist.id = NULL}.}

\item{verbose}{Logical indicating whether progress messages are emitted.
Default: \code{verbose = TRUE}.}
}
\value{
A tibble with column \code{INDIVIDUALS}.
}
\description{
Read a file or object with blacklisted individuals.
Used internally in \href{https://github.com/thierrygosselin/genometranslator}{genometranslator}
and might be of interest for users.
}
\section{Dependencies}{

Required package dependencies are declared in \code{DESCRIPTION} and are
installed with \pkg{genometranslator}. Any additional dependency needed only
for this format or option is identified in this help page. Use
\code{genometranslator_dependencies()} to inspect the availability of core
packages, optional packages, and external executables.
}

\examples{
\dontrun{
bl <- genometranslator::read_blacklist_id("blacklist.tsv")
}
}
