% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genind.R
\name{read_genind}
\alias{read_genind}
\title{Read a genind object to a GDS or tidy dataframe}
\usage{
read_genind(data, tidy = TRUE, gds = TRUE, write = FALSE, verbose = FALSE)
}
\arguments{
\item{data}{(path or object) A genind object in the global environment or
path to a genind file that will be open with \code{readRDS}.}

\item{tidy}{(logical) Generate a tidy dataset.
Default: \code{tidy = TRUE}.}

\item{gds}{(optional, logical) To write a radiator gds object.
Currently, for biallelic datasets only.
Default: \code{gds = TRUE}.}

\item{write}{(optional, logical) To write in the working directory the tidy
data. The file is written with \code{radiator_genind_DATE@TIME.rad}.
Default: \code{write = FALSE}.}

\item{verbose}{Logical indicating whether progress messages are emitted.
Default: \code{verbose = FALSE}.}
}
\description{
Read a genind object or file from
\href{https://github.com/thibautjombart/adegenet}{adegenet} to a tidy dataframe.
Used internally in \href{https://github.com/thierrygosselin/genometranslator}{genometranslator}
and might be of interest for users.
}
\note{
\href{https://github.com/thibautjombart/adegenet}{genind} objects, like
genepop, are not optimal genomic format for RADseq datasets,
they lack important genotypes and markers metadata: chromosome, locus, snp,
position, read depth, allele depth, etc.
\href{https://github.com/thibautjombart/adegenet}{genlight} object is a more
interesting container and is memory efficient, see \code{\link{read_genlight}}.

By default allele names will be kept for the tidy dataset,
if the alleles is numeric and length < 3.

In the unlikely event that the genind object as no stratification/population,
\emph{pop} will be added to the strata column.
}
\section{Dependencies}{

Required package dependencies are declared in DESCRIPTION and installed
with genometranslator. Run \code{genometranslator_dependencies()} to inspect
core packages, optional packages, and external executables.

Reading and writing \code{genind} objects requires the optional CRAN package
\href{https://cran.r-project.org/package=adegenet}{\pkg{adegenet}}.
}

\references{
Jombart T (2008) adegenet: a R package for the multivariate
analysis of genetic markers. Bioinformatics, 24, 1403-1405.

Jombart T, Ahmed I (2011) adegenet 1.3-1:
new tools for the analysis of genome-wide SNP data.
Bioinformatics, 27, 3070-3071.
}
\author{
Thierry Gosselin \email{thierrygosselin@icloud.com}
}
