% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genlight.R
\name{read_genlight}
\alias{read_genlight}
\title{Read a genlight object into a tidy data frame and/or GDS object/file}
\usage{
read_genlight(
  data,
  tidy = TRUE,
  gds = TRUE,
  write = FALSE,
  verbose = FALSE,
  parallel.core = parallel::detectCores() - 1
)
}
\arguments{
\item{data}{(path or object) A genlight object in the global environment or
path to a genlight file that will be open with \code{readRDS}.}

\item{tidy}{(logical) Generate a tidy dataset.
Default: \code{tidy = TRUE}.}

\item{gds}{(optional, logical) To write a radiator gds object.
Default: \code{gds = TRUE}.}

\item{write}{(optional, logical) To write in the working directory the tidy
data. The file is written with \code{radiator_genlight_DATE@TIME.rad}.
Default: \code{write = FALSE}.}

\item{verbose}{Logical indicating whether progress messages are emitted.
Default: \code{verbose = FALSE}.}

\item{parallel.core}{Number of workers available for parallel operations.
Default: \code{parallel.core = parallel::detectCores() - 1}.}
}
\description{
Read a genlight object into a tidy data frame and/or GDS object/file.
Used internally in \href{https://github.com/thierrygosselin/genometranslator}{genometranslator}
and might be of interest for users.
}
\details{
A string of the same dimension is generated when genlight:
\enumerate{
\item \code{is.null(genlight@pop)}: pop will be integrated
in the tidy dataset.
\item \code{is.null(data@chromosome)}: CHROM1 will be integrated
in the tidy dataset.
\item \code{is.null(data@loc.names)}: LOCUS1 to \code{ncol(genlight)}
will be integrated in the tidy dataset.
\item \code{is.null(data@position)}: an integer string of
length = \code{ncol(genlight)} will be integrated in the tidy dataset.
}

\strong{Note: that if all CHROM, LOCUS and POS is missing the function will be terminated}
}
\section{Dependencies}{

Required package dependencies are declared in DESCRIPTION and installed
with genometranslator. Run \code{genometranslator_dependencies()} to inspect
core packages, optional packages, and external executables.

Reading and writing \code{genlight} objects requires the optional CRAN
package \href{https://cran.r-project.org/package=adegenet}{\pkg{adegenet}}.
}

\references{
Jombart T (2008) adegenet: a R package for the multivariate
analysis of genetic markers. Bioinformatics, 24, 1403-1405.

Jombart T, Ahmed I (2011) adegenet 1.3-1:
new tools for the analysis of genome-wide SNP data.
Bioinformatics, 27, 3070-3071.
}
\author{
Thierry Gosselin \email{thierrygosselin@icloud.com}
}
