% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/marker_metadata.R
\name{separate_markers}
\alias{separate_markers}
\title{Separate markers column into chrom, locus and pos}
\usage{
separate_markers(
  data,
  sep = "__",
  markers.meta.lists.only = FALSE,
  markers.meta.all.only = FALSE,
  generate.markers.metadata = TRUE,
  generate.ref.alt = FALSE,
  biallelic = NULL,
  parallel.core = parallel::detectCores() - 1,
  verbose = TRUE
)
}
\arguments{
\item{data}{An object with a column named \code{MARKERS}.
If \code{CHROM}, \code{LOCUS}, \code{POS} are already present, the function
returns the dataset untouched.
The data can be whitelists and blacklists of markers or tidy datasets or
genome GDS object.}

\item{sep}{(optional, character) Separator used to identify the different
field in the \code{MARKERS} column.

When the \code{MARKERS} column doesn't have separator and the function is used
to generate markers metadata column:
\code{"CHROM", "LOCUS", "POS", "REF", "ALT"}, use \code{sep = NULL}.
Default: \code{sep = "__"}.}

\item{markers.meta.lists.only}{(logical, optional)
Allows to keep only the markers metadata:
\code{"VARIANT_ID", "MARKERS", "CHROM", "LOCUS", "POS"}, useful for whitelist
or blacklist.
Default: \code{markers.meta.lists.only = FALSE}.}

\item{markers.meta.all.only}{(logica, optionall)
Allows to keep all available markers metadata:
\code{"VARIANT_ID", "MARKERS", "CHROM", "LOCUS", "POS", "COL", "REF", "ALT"},
useful inside genomic translation workflows.
Default: \code{markers.meta.all.only = FALSE}.}

\item{generate.markers.metadata}{(logical, optional)
Generate missing markers metadata when missing.
\code{"CHROM", "LOCUS", "POS"}.

Default: \code{generate.markers.metadata = TRUE}.}

\item{generate.ref.alt}{(logical, optional) Generate missing REF/ALT alleles
with: REF = A and ALT = C (for biallelic datasets, only).
It is turned off automatically
when argument \code{markers.meta.lists.only = TRUE} and
on automatically when argument \code{markers.meta.all.only = TRUE}
Default: \code{generate.ref.alt = FALSE}.}

\item{biallelic}{(logical) Speed up the function execution by entering
if the dataset is biallelic or not. Used internally for verification, before
generating REF/ALT info.
The argument is required when \code{generate.ref.alt = TRUE} and REF/ALT
information is absent. Biallelic detection is analysis policy and is not
performed automatically by this package.
Default: \code{biallelic = NULL}.}

\item{parallel.core}{Default: \code{parallel.core = parallel::detectCores() - 1}.}
}
\value{
The same data in the global environment, with 3 new columns:
\code{CHROM}, \code{LOCUS}, \code{POS}. Additionnal columns may be genrated,
see arguments documentation.
}
\description{
Genometranslator uses unique marker names by combining
\code{CHROM}, \code{LOCUS}, \code{POS} columns, with double underscore
separators, into \code{MARKERS = CHROM__LOCUS__POS}.

Used internally in \code{genometranslator}
and might be of interest for users who need to get back to the original metadata the
function provides an easy way to do it.
}
\examples{
\dontrun{
whitelist <- genometranslator::separate_markers(data = whitelist.markers)
tidy.data <- genometranslator::separate_markers(data = bluefintuna.data)
}
}
\seealso{
\code{\link{generate_markers_metadata}}
}
\author{
Thierry Gosselin \email{thierrygosselin@icloud.com}
}
