% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/strata.R
\name{summary_strata}
\alias{summary_strata}
\title{Summary of strata}
\usage{
summary_strata(strata)
}
\arguments{
\item{strata}{(path or object) The strata file or object.}
}
\value{
\enumerate{
\item Number of strata/populations
\item Number of individuals
\item Number of individuals per populations
\item Number of duplicate ids.
}
}
\description{
Summarise the information of a strata file or object.
Used internally in \href{https://github.com/thierrygosselin/genometranslator}{genometranslator}
and might be of interest for users.
}
\examples{
\dontrun{
genometranslator::summary_strata(strata)
}
}
\seealso{
\code{\link{read_strata}},
\code{\link{change_pop_names}},
\code{\link{join_strata}}
}
