% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plink.R
\name{tidy_plink}
\alias{tidy_plink}
\title{Tidy PLINK tped and bed files}
\usage{
tidy_plink(
  data,
  parallel.core = parallel::detectCores() - 1,
  verbose = FALSE,
  ...
)
}
\arguments{
\item{data}{The PLINK file.
\itemize{
\item {bi-allelic data only}. For haplotypes use VCF.
\item \code{tped} file format: the corresponding \code{tfam} file must be in the directory.
\item \code{bed} file format: IS THE PREFERRED format, the corresponding
\code{fam} and \code{bim} files must be in the directory.
}}

\item{parallel.core}{Number of workers available for parallel operations.
Default: \code{parallel.core = parallel::detectCores() - 1}.}

\item{verbose}{Logical indicating whether progress messages are emitted.
Default: \code{verbose = FALSE}.}

\item{...}{Additional arguments passed to lower-level readers, translators,
or writers.}
}
\value{
A tidy tibble of the PLINK file.
}
\description{
Transform bi-allelic PLINK files in \code{.tped} or \code{.bed} formats
into a tidy dataset.

Used internally in \href{https://github.com/thierrygosselin/genometranslator}{genometranslator}
and \href{https://github.com/thierrygosselin/assigner}{assigner}
and might be of interest for users.
Ensure marker type, missingness, allele coding, and any linkage pruning are
appropriate for the intended PLINK analysis before export.
}
\section{Advance mode}{


\emph{dots-dots-dots ...} allows to pass several arguments for fine-tuning the function:
\enumerate{
\item \code{calibrate.alleles}: logical. For \code{tped} files, if
\code{calibrate.alleles = FALSE} the function runs faster
but REF/ALT alleles may not be calibrated. The default assumes the users or
sotware producing the PLINK file calibrated the alleles.
Default: \code{calibrate.alleles = FALSE}.
}
}

\examples{
\dontrun{
data <- genometranslator::tidy_plink(data = "my_plink_file.bed", verbose = TRUE)


# when conversion is required from TPED to BED, in Terminal:
# plink --tfile my_plink_file --make-bed --allow-no-sex --allow-extra-chr --chr-set 95
}
}
\references{
Zheng X, Gogarten S, Lawrence M, Stilp A, Conomos M, Weir BS,
Laurie C, Levine D (2017). SeqArray -- A storage-efficient high-performance
data format for WGS variant calls.
Bioinformatics.

PLINK: a tool set for whole-genome association and population-based linkage
analyses.
American Journal of Human Genetics. 2007: 81: 559–575. doi:10.1086/519795
}
\seealso{
\href{https://www.cog-genomics.org/plink/1.9/}{PLINK}

\code{\link{read_plink}}
}
\author{
Thierry Gosselin \email{thierrygosselin@icloud.com}
}
