% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_betadiv.R
\name{write_betadiv}
\alias{write_betadiv}
\title{Write a betadiv file from a tidy data frame}
\usage{
write_betadiv(data)
}
\arguments{
\item{data}{A tidy data frame object in the global environment or
a tidy data frame in wide or long format in the working directory.
\emph{How to get a tidy data frame ?}
Look into \pkg{genometranslator} \code{\link{tidy_genome}}.}
}
\value{
A betadiv object is returned.
}
\description{
Write a betadiv file from a tidy data frame.
Used internally in \href{https://github.com/thierrygosselin/genometranslator}{genometranslator}
and \href{https://github.com/thierrygosselin/assigner}{assigner}
and might be of interest for users.
}
\section{Data filtering}{

This writer does not silently filter markers or individuals. It may validate
requirements imposed by the destination format and stop with an informative
error when the input is unsuitable. It is the user's responsibility to
filter and quality-control the data appropriately for the intended analysis
before generating the output. Use
\href{https://thierrygosselin.github.io/radr/}{\pkg{radr}} or another
suitable workflow when filtering is required.
}

\section{Dependencies}{

Required package dependencies are declared in \code{DESCRIPTION} and are
installed with \pkg{genometranslator}. Any additional dependency needed only
for this format or option is identified in this help page. Use
\code{genometranslator_dependencies()} to inspect the availability of core
packages, optional packages, and external executables.
}

\references{
Lamy T, Legendre P, Chancerelle Y, Siu G, Claudet J (2015)
Understanding the Spatio-Temporal Response of Coral Reef Fish Communities to
Natural Disturbances: Insights from Beta-Diversity Decomposition.
PLoS ONE, 10, e0138696.
}
\seealso{
\code{beta.div} is available on Pierre Legendre web site \url{http://adn.biol.umontreal.ca/~numericalecology/Rcode/}
}
\author{
Laura Benestan \email{laura.benestan@icloud.com} and
Thierry Gosselin \email{thierrygosselin@icloud.com}
}
